backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
pydantic==2.4.2
requests==2.31.0
python-dotenv==1.0.0

# Document processing
PyPDF2==3.0.1
pdfplumber==0.10.2
pdf2image==1.16.3
pytesseract==0.3.10
python-docx==0.8.11
odfpy==1.4.1
pandas==2.1.1
openpyxl==3.1.2
beautifulsoup4==4.12.2
html2text==2020.1.16
Pillow==10.0.1

# NLP and AI
google-generativeai==0.3.1
openai==1.3.5
anthropic==0.5.0
mistralai==0.0.7

# Utilities
aiofiles==23.2.1
python-jose==3.3.0
tenacity==8.2.3
